Always let an in-flow legend be part of any flow thread parent.
The containing block of a rendered legend of a fieldset that also
establishes a multicol container is the fieldset itself, not the flow
thread (and also not the anonymous fieldset contents wrapper).

However, since dynamic changes may change which legend is the rendered
legend, we'll keep it simple, and always let them be treated as part of
the flow thread - so that they trigger creation of a
LayoutMultiColumnSet - so that we are guaranteed to actually have a
LayoutMultiColumnSet if the rendered legend ceases to be the rendered
one.

Bug: 1122722
Change-Id: I5a287e367f8824dc0c5a27620c74d2ec6eedfd34