Transferable Streams: Fix crash on deserialization error
The transferable streams implementation would crash when it received a
chunk that wasn't deserializable, because it failed to enter a
ScriptScope before creating a V8 object.

Fix it.

Also fix the message on the DOMException that is created when
deserialization fails.

Add a test for deserialization failure. This is hard to trigger in a
cross-browser fashion. In this test, we use a WASM module, which can't
be transferred to a different site.

BUG=1122725

Change-Id: I6e49c69978388357a3c6006cc02dce8f0a949954